crates/gc: impl new gc
This GC is based on the rust-gc crate. It uses the Trace traits root
and unroot method to keep track of what is rooted, and reference
counting to keep track of how many roots are currently referencing an
allocation. It traces anything that has a root reference and uses mark
and sweep to determine what is safe to free.