crates/gc: add debug assert guards on root and unroot
This also removes the root and unroot inner.data call in Gc::root and
Gc::unroot. Each Gc gets unrooted when it gets put in another Gc, so
there is no need to call it recursively for all of the parent
Gc's. Without removing this, test cases fail due to double unrooting.