crates/reader: rewrite, new SEXpr enum that retains span info
This rewrites the reader, and changes from using nom to logos for
parsing.

The main change is that we introduce a new SExpr type, which is
similar to the old Value type, but lists are now Vec's instead of
conses, and the enum variants hold a Range<usize> which is their
position in the source file they were parsed from.

Quoted forms like ('a) will have their span set to start and end at
the position of the quote symbol, this way error messages for macros
will point to the correct place.