crates/compiler: only store struct type parameters in type env
This change makes the struct TypeInfo and Type structs store the
type parameters for the struct instead of all of the fields. The type
parameters is the only info that is actually required to be stored to
type check, and it allows more easily working with struct type
annotations.

This change also changes how type annotations are parsed. We store a
hashmap of the names of user defined types, and check all annotations
to see if they are in the hashmap to determine whether they are a user
defined struct or variant.