crates/compiler: pop scope after compiling macro body
We forgot to pop the scope that is created for compiling the macros
body. This caused the macros scopes to be in effect for the rest of
the compilation.

This caused the the compiler to create upvalues for
top-level functions, which caused the VM to panic, for variables that
weren't "in scope" and happened to have the same name as one of the
variables in the leftover scopes that didn't get popped.